mem_sim
A memory simulator written in C. It simulates how memory is accessed and overwritten. The program is run with 3 parameters: replacement policy, associativity and number of cache blocks.
Replacement policy is how memory is replaced (i.e. which address it is placed in in the cache block) and can be FIFO (first in first out), LRU (Least Recently Used) and Random.
Associativity is how wide cache blocks are, and can be any power of 2 up to the number of cache blocks.
Number of cache blocks can be 16, 64, 256 or 1024.